chore: refacto migration function
